+++ src/Responses/PaginatedResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SignVault\Responses;
+namespace Signori\Responses;
 
 /**
  * Wraps a paginated list response.