+++ src/Responses/SignerResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SignVault\Responses;
+namespace Signori\Responses;
 
 final class SignerResponse extends BaseResponse
 {